修正: 字符串应接收常量引用进行转换而非引用参数局部对象;; 新增: container_sink 用于直接向 STL 容器流输入;

+++ test/strconv.cpp
@@ -12,5 +12,5 @@ int strconv_test(int argc, char* argv[]) {
     LOGGER() << "\t" << strconv::str2ul(o) << std::endl;
     LOGGER() << "\t" << strconv::str2ui(s) << std::endl;
     LOGGER() << "\t" << strconv::str2us(v) << std::endl;
-    LOGGER() << "\t" << strconv::str2uc(a) << std::endl;
+    LOGGER() << "\t" << (int)strconv::str2uc(a) << std::endl;
 }
